+++ src/main/java/control/IField.java
@@ -8,4 +8,6 @@ public interface IField {
     public void setFigureOnField(int x, int y, Figure figure);
 
     public boolean isFieldUnitEmpty(int x, int y);
+
+    public Figure getFigureFromPosition(final int x, final int y);
 }
