+++ src/main/java/view/ConsoleRead.java
@@ -1,4 +1,4 @@
-package control;
+package view;
 
 import java.io.BufferedReader;
 import java.io.IOException;
