--- src/main/java/view/IPrintScreen.java
@@ -1,6 +1,0 @@
-package view;
-
-public interface IPrintScreen {
-
-    void printScreen();
-}
