--- src/main/java/view/PrintScreen.java
@@ -1,8 +1,0 @@
-package view;
-
-public class PrintScreen implements IPrintScreen {
-
-    public void printScreen() {
-
-    }
-}
